wip: separated contract-generate. export StringContractDefinition. Allow --contract option
