+++ packages/libs/pactjs-generator/src/index.ts
@@ -6,6 +6,16 @@ export * from './contract/generation/FileContractDefinition';
 /**
  * @alpha
  */
+export * from './contract/generation/StringContractDefinition';
+
+/**
+ * @alpha
+ */
+export * from './contract/generation/IContractDefinition';
+
+/**
+ * @alpha
+ */
 export * from './contract/generation/generator';
 
 /**
